+++ src/main/java/de/uib/configed/gui/NewClientDialog.java
@@ -494,7 +494,7 @@ public final class NewClientDialog extends FGeneralDialog {
 		FSelectionList groupsSelectionDialog = new FSelectionList(ConfigedMain.getMainFrame(),
 				Configed.getResourceValue("NewClientDialog.groupSelectionDialog.title"), true,
 				new String[] { Configed.getResourceValue("buttonCancel"), Configed.getResourceValue("buttonOK") }, 500,
-				300, "groupselection");
+				300);
 		groupsSelectionDialog.enableMultiSelection();
 		groupsSelectionDialog.setListData(
 				PersistenceControllerFactory.getPersistenceController().getGroupDataService().getHostGroupIds());