+++ src/main/java/de/uib/configed/gui/SavedSearchesDialog.java
@@ -87,8 +87,8 @@ public class SavedSearchesDialog extends FEditStringList {
 
 		SearchTargetModel searchTargetModel = new SearchTargetModelFromJList(visibleList, new ArrayList<>(),
 				new ArrayList<>());
-		searchPane = new TableSearchPane(searchTargetModel, "savedsearches");
-		searchPane.setSearchMode(TableSearchPane.FULL_TEXT_SEARCH);
+		searchPane = new TableSearchPane(searchTargetModel);
+		searchPane.setSearchMode(TableSearchPane.SearchMode.FULL_TEXT_SEARCH);
 		searchPane.setNarrow(true);
 		editingArea.add(searchPane, BorderLayout.NORTH);