+++ CHANGELOG.md
@@ -3,6 +3,13 @@
 ## Fixed
 - When group, that has subgroups, is permitted for a user, the subgroups aren't included in a group.
 
+## Changed
+- Remove config states usage in `TableSearchPane`
+- Remove usage of icons in narrow layout of `TableSearchPane` (use checkboxes instead)
+
+## Added
+- New option for searching multiple words in `TableSearchPane`
+
 ## [4.3.2.6] - 2024-03-06
 ## Fixed
 - `Nullpointerexception` on reload